2-floor.py

+++ 0x00-python_variable_annotations/2-floor.py
@@ -5,4 +5,4 @@ import math
 
 def floor(n: float) -> float:
     """ Functions return floor of a float """
-    return float(math.floor(n))
+    return math.floor(n)
